monitoring-system/feature-engineer/requirements.txt
kafka-python
kubernetes
requests